Add the second css exercise
Here i add the final version for the moment of
this exercise, i believe that is similar to what
i need to do in the original screenshot.

i do it because because what is more sad that
not have nothing in this marvelous folder?
So i make the wonderful decision of do this
exercise and i think that their goal with it is
use a lot more the id and class selector.